move gpt2 pipeline forwards to modeling folder
